External profiler + fixes (#36)
* Add Tracy profiler
* Add glfw to enable profiler building on windows
* Fix EASTL bug
* Replace List with ListPOD in a few places.
* Fix a few windows platform bugs + restore the original order of
importers.
* Fix another string_view issue + replace List/Vector in a few places.
* Use ConfigFile support in ResourceFormatImporter::_get_path_and_type
* Possible fix for vs2017 compiler running out of heap during
rasterizer_canvas_gles3.cpp compilation.

+++ core/se_string.h
@@ -17,12 +17,12 @@ extern const se_string null_se_string; // used to return 'null' string reference
 
 template<>
 struct Hasher<se_string> {
-    _FORCE_INLINE_ uint32_t operator()(const se_string &s) const { return eastl::string_hash<se_string>()(s); }
+    _FORCE_INLINE_ uint32_t operator()(const se_string &s) const { return uint32_t(eastl::string_hash<se_string>()(s)); }
 };
 namespace StringUtils {
 
 static inline uint32_t hash(se_string_view sv) {
-    return eastl::hash<se_string_view>()(sv);
+    return uint32_t(eastl::hash<se_string_view>()(sv));
 }
 }
 namespace eastl {
